+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 x.x.x Release notes (yyyy-MM-dd)
 =============================================================
 ### Enhancements
-* Added new method `Realm.createOrUpdate(type, values, mode)`. ($)
-* Added new update mode , which will only update values that actually changed. 
+* Added a new `Realm.UpdateMode` enum with the values: `never`, `modified`, `all`. This replaces the current 
+  `Realm.create(type, properties, update)` with `Realm.create(type, properties, updateMode)`.
+  `Realm.create(type, properties, 'modified')` is a new mode that only update existing properties that actually
+  changed, while `Realm.create(type, properties, 'never')` is equal to `Realm.create(type, properties, false)` and
+  `Realm.create(type, properties, 'all')` is equal to `Realm.create(type, properties, false)`. 
+  `Realm.create(type, properties, update)` is now deprecated. ([#2089](https://github.com/realm/realm-js/issues/2089))
 
 ### Fixed
 * <How to hit and notice issue? what was the impact?> ([#????](https://github.com/realm/realm-js/issues/????), since v?.?.?)